added client_wrapper.py
